Restrict access to Assignments and Uploads
Gracefully deal with Forbidden (403) errors on the frontend

+++ services/src/main/java/org/openlearn/web/rest/AssignmentResource.java
@@ -2,9 +2,11 @@ package org.openlearn.web.rest;
 
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 import io.swagger.annotations.ApiParam;
+import org.openlearn.domain.Assignment;
 import org.openlearn.domain.FileInformation;
 import org.openlearn.dto.AssignmentDTO;
 import org.openlearn.security.AuthoritiesConstants;
+import org.openlearn.security.SecurityUtils;
 import org.openlearn.service.AssignmentService;
 import org.openlearn.service.CourseService;
 import org.openlearn.service.StorageService;
@@ -102,8 +104,12 @@ public class AssignmentResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity create(@RequestBody @Valid final AssignmentDTO assignmentDTO) throws URISyntaxException {
 		log.debug("POST request to create assignment : {}", assignmentDTO);
-		AssignmentDTO response = assignmentService.save(assignmentDTO);
-		return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		if (hasCreateUpdateDeleteAuthority(assignmentDTO)) {
+			AssignmentDTO response = assignmentService.save(assignmentDTO);
+			return ResponseEntity.created(new URI(ENDPOINT + response.getId())).body(response);
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	/**
@@ -117,8 +123,12 @@ public class AssignmentResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity update(@RequestBody @Valid final AssignmentDTO assignmentDTO) {
 		log.debug("PUT request to update assignment : {}", assignmentDTO);
-		AssignmentDTO response = assignmentService.save(assignmentDTO);
-		return ResponseEntity.ok(response);
+		if (hasCreateUpdateDeleteAuthority(assignmentDTO)) {
+			AssignmentDTO response = assignmentService.save(assignmentDTO);
+			return ResponseEntity.ok(response);
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	/**
@@ -132,8 +142,12 @@ public class AssignmentResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR})
 	public ResponseEntity delete(@PathVariable final Long id) {
 		log.debug("DELETE request to delete assignment : {}", id);
-		assignmentService.delete(id);
-		return ResponseEntity.ok().build();
+		if (hasCreateUpdateDeleteAuthority(assignmentService.findOne(id))) {
+			assignmentService.delete(id);
+			return ResponseEntity.ok().build();
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 
@@ -150,20 +164,28 @@ public class AssignmentResource {
 	public ResponseEntity uploadCourseFile(@PathVariable final Long assignmentId,
 								   @RequestParam("file") MultipartFile file,
 								   RedirectAttributes redirectAttributes) throws URISyntaxException {
-		FileInformation response = storageService.store(file, assignmentId, null);
-		redirectAttributes.addFlashAttribute("message",
-			"You successfully uploaded " + file.getOriginalFilename() + "!");
+		if (canUploadFilesToAssignment(assignmentService.findOne(assignmentId))) {
+			FileInformation response = storageService.store(file, assignmentId, null);
+			redirectAttributes.addFlashAttribute("message",
+				"You successfully uploaded " + file.getOriginalFilename() + "!");
 
-		URI location = new URI(ENDPOINT + response.getId());
-		return ResponseEntity.created(location).body(response);
+			URI location = new URI(ENDPOINT + response.getId());
+			return ResponseEntity.created(location).body(response);
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	@GetMapping(path="/{assignmentId}/uploads")
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity getUploads(@PathVariable final Long assignmentId) {
 		log.debug("GET request to get course uploads for assignment " + assignmentId);
-		List<S3ObjectSummary> response = storageService.getUploads(assignmentId, null).getObjectSummaries();
-		return ResponseEntity.ok(response);
+		if (canUploadFilesToAssignment(assignmentService.findOne(assignmentId))) {
+			List<S3ObjectSummary> response = storageService.getUploads(assignmentId, null).getObjectSummaries();
+			return ResponseEntity.ok(response);
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
+		}
 	}
 
 	@GetMapping(path="/{assignmentId}/upload/{keyName:.+}")
@@ -171,22 +193,26 @@ public class AssignmentResource {
 	public ResponseEntity getUpload(@PathVariable final Long assignmentId,
 									@PathVariable final String keyName) {
 		log.debug("GET request to get course upload : {}", keyName);
-		InputStream response = storageService.getUpload(assignmentId, null, keyName);
-		if (response != null) {
-			try {
-				byte[] out = org.apache.commons.io.IOUtils.toByteArray(response);
+		if (hasCreateUpdateDeleteAuthority(assignmentService.findOne(assignmentId))) {
+			InputStream response = storageService.getUpload(assignmentId, null, keyName);
+			if (response != null) {
+				try {
+					byte[] out = org.apache.commons.io.IOUtils.toByteArray(response);
 
-				HttpHeaders responseHeaders = new HttpHeaders();
-				responseHeaders.add("content-disposition", "attachment; filename=" + keyName);
-				// responseHeaders.add("Content-Type", type);
+					HttpHeaders responseHeaders = new HttpHeaders();
+					responseHeaders.add("content-disposition", "attachment; filename=" + keyName);
+					// responseHeaders.add("Content-Type", type);
 
-				return new ResponseEntity(out, responseHeaders, HttpStatus.OK);
-			} catch (Exception e) {
-				new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+					return new ResponseEntity(out, responseHeaders, HttpStatus.OK);
+				} catch (Exception e) {
+					new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+				}
+				return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
+			} else {
+				return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
 			}
-			return new ResponseEntity("File Not Found", HttpStatus.NOT_FOUND);
 		} else {
-			return new ResponseEntity ("File Not Found", HttpStatus.NOT_FOUND);
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
 		}
 	}
 
@@ -194,11 +220,42 @@ public class AssignmentResource {
 	@Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.ORG_ADMIN, AuthoritiesConstants.INSTRUCTOR, AuthoritiesConstants.STUDENT})
 	public ResponseEntity deleteUpload(@PathVariable final Long assignmentId,
 									   @PathVariable final String keyName) {
-		try {
-			storageService.deleteUpload(assignmentId, null, keyName);
-		} catch (Exception e) {
-			return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+		if (hasCreateUpdateDeleteAuthority(assignmentService.findOne(assignmentId))) {
+			try {
+				storageService.deleteUpload(assignmentId, null, keyName);
+			} catch (Exception e) {
+				return new ResponseEntity("Error deleting upload", HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			return new ResponseEntity(HttpStatus.OK);
+		} else {
+			return new ResponseEntity(HttpStatus.FORBIDDEN);
 		}
-		return new ResponseEntity(HttpStatus.OK);
+	}
+
+	private Boolean canUploadFilesToAssignment(AssignmentDTO assignmentDTO) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(assignmentDTO) ||
+			isCourseInstructor(assignmentDTO) ||
+			isCourseStudent(assignmentDTO);
+	}
+
+	private Boolean hasCreateUpdateDeleteAuthority(AssignmentDTO assignmentDTO) {
+		return SecurityUtils.isAdmin() ||
+			isOrgAdmin(assignmentDTO) ||
+			isCourseInstructor(assignmentDTO);
+	}
+
+	private Boolean isOrgAdmin(AssignmentDTO assignmentDTO) {
+		return SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ORG_ADMIN) &&
+			assignmentService.inOrgOfCurrentUser(assignmentDTO);
+
+	}
+
+	private Boolean isCourseInstructor(AssignmentDTO assignmentDTO) {
+		return assignmentService.currentUserIsCourseInstructor(assignmentDTO);
+	}
+
+	private Boolean isCourseStudent(AssignmentDTO assignmentDTO) {
+		return assignmentService.currentUserIsEnrolledIn(assignmentDTO);
 	}
 }
